// src/models/management/core/operation/operationCostingDetail.interface.ts
type BaseDetailType = {
  id: number;
  costingId: number;
  supplierId: number;
  operationId: number;
  sellableId: number;
  sellableCode: string;
  paymentQuantity: number;
  amount: number;
  totalAmount: number;
  isLocked: boolean;
  lockedUpdate: string | null;
  isFinalized: boolean;
  finalizedUpdate: string | null;
  status: "NEW" | "DONE";
};
export type AirDetailType = {
  type: EStockType.AIRTICKET | EStockType.INSURANCE | EStockType.OTHER;
  details: {
    classOfService: string;
    tripType: "ONEWAY" | "ROUNDTRIP" | "MULTICITY";
    adult: number;
    child: number;
    infant: number;
    fullItinerary: string;
    departureDate: string;
    arrivalDate: string;
    specialRequest: string;
    remark: string;
  } | null;
};
export type VisaDetailType = {
  type: EStockType.VISASERVICES | EStockType.OTHER;
  details: {
    specialRequest: string;
    remark: string;
  } | null;
};
export type LandPackageDetailType = {
  type: EStockType.PACKAGE | EStockType.TOURPACKAGE | EStockType.GUIDE | EStockType.OTHER;
  details: {
    specialRequest: string;
    remark: string;
  } | null;
};
export type GuideDetailType = {
  type: EStockType.OTHER;
  details: {
    guideType: "DOMESTIC" | "OUTBOUND" | "INBOUND";
    destination: string;
    quantity: number;
    specialRequest: string;
    remark: string;
  } | null;
};
export type HotelDetailType = {
  type: EStockType.ROOM | EStockType.OTHER;
  details: {
    roomType: RoomingType;
    checkIn: string;
    checkOut: string;
    quantity: number;
    specialRequest: string;
    remark: string;
  } | null;
};
export type RestaurantDetailType = {
  type: EStockType.TABLE | EStockType.OTHER;
  details: {
    guideType: "DOMESTIC" | "OUTBOUND" | "INBOUND";
    destination: string;
    quantity: number;
    specialRequest: string;
    remark: string;
  } | null;
};
export type TransportDetailType = {
  type: EStockType.CRUISE | EStockType.VEHICLE | EStockType.OTHER;
  details: {
    pickUpDate: string;
    pickUpLocation: string;
    dropOffLocation: string;
    quantity: number;
    carType: string;
    carModel: string;
    specialRequest: string;
    remark: string;
  } | null;
};
export type InsuranceDetailType = {
  type: EStockType.OTHER;
  details: {
    specialRequest: string;
    remark: string;
  } | null;
};

export type ICostingDetailItem<T = void> = T & BaseDetailType;

export interface IOperationCostingDetail {
  id: number;
  type: EInventoryType;
  supplierId: number;
  operationId: number;
  sellableId: number;
  sellableCode: string;
  totalCost: number;
  status: "NEW" | "DONE";
  sellableMinimal: Pick<
    ISellable,
    | "recId"
    | "sellableTemplateId"
    | "type"
    | "code"
    | "cap"
    | "open"
    | "used"
    | "available"
    | "closeDate"
    | "validFrom"
    | "validTo"
    | "startDate"
    | "endDate"
  > & {
    sellableTemplateCode: null;
    configs: null;
    sellableDetails: null;
    promotions: null;
  };
  templateMinimal: Pick<ITemplateSellable, "recId" | "cmsIdentity" | "type" | "code" | "name" | "inventoryTypeList"> & {
    cms: null;
    sellables: null;
    cmsMustKnow: null;
  };
  details: ICostingDetailItem<
    | AirDetailType
    | RestaurantDetailType
    | InsuranceDetailType
    | VisaDetailType
    | GuideDetailType
    | HotelDetailType
    | LandPackageDetailType
    | TransportDetailType
  >[];
}